address Adriano's comments
